Netevent + Zlib library
The Zlib library allows one to compress and decompress the payload,
since initially we were trying to implement a full Doom player that
would interact with the host using the Networking protocol.

However, this plan was ditched after checking that we would have to
implement packet tracking, since IIRC it's the responsibility of the
receiver to track if there was some missing packets, and it had to
increment and return the sequence number provided by the host in the
first place.

The Netevent module is just a wrapper around Zlib and :gen_udp to make
it easy to send Netevent messages to the game.